refactored module code
