+++ src/napari_roxas_ai/_preparation/__init__.py
@@ -1,3 +1,5 @@
+from ._file_overwrite_dialog import FileOverwriteDialog
 from ._preparation_widget import PreparationWidget
+from ._worker import Worker
 
-__all__ = ("PreparationWidget",)
+__all__ = ("PreparationWidget", "FileOverwriteDialog", "Worker")
